+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All important changes in vue-frontend-widgets
 
+## [0.1.6] - 2024-12-17
+Intersection table:
+- Emitting events with the columns updates
+- Emitting events with the rows updates
+- Fixed problem with the loader position
+
 ## [0.1.1] - 2024-11-20
 - Added data-test-id attribute to select options and KPI card values to ease acceptance testing